cleaned up some files
changed archival to switch between two different method (one for compressed files, and one for uncompressed). This is because we were hitting resource limits for very large files which we could not compress. for now the solution will be to simply not compress files larger than one gigabyte and instead stream them into the archive instead.

this code is hell, maybe ome day I will make it better :)